Updates to fix spring decoder issue
